functions that prints the first 50 Fibonacci numbers

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -8,7 +8,7 @@
 int main(void)
 {
 	int i;
-	unsigned long long int a, b, sum;
+	unsigned long int a, b, sum;
 
 	a = 0;
 	b = 1;
@@ -16,7 +16,7 @@ int main(void)
 	for (i = 0; i < 50; i++)
 	{
 		sum = a + b;
-		printf("%11llu, ", sum);
+		printf("%11d, ", sum);
 
 		a = b;
 		b = sum;
